pyproject.toml
[tool.poetry]
name = "desafio-dio-integrando-python-com-sqlite-e-mongodb"
version = "0.1.0"
description = "Usando as bibliotecas sqlAlchemy e pymongo para integração com banco de dados"
authors = ["Denilson Silva <162153089+AroMight@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
SQLAlchemy = "^2.0.30"
pymongo = "^4.7.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
